+++ packages/cubejs-playground/src/components/ChartRenderer/ChartRenderer.tsx
@@ -8,6 +8,8 @@ import type { PivotConfig, Query, ChartType } from '@cubejs-client/core';
 
 import { Button, CubeLoader } from '../../atoms';
 import { UIFramework } from '../../types';
+import { event } from '../../events';
+import { QueryStatus } from '../../PlaygroundQueryBuilder';
 
 const { Text } = Typography;
 
@@ -68,7 +70,7 @@ export type TQueryLoadResult = {
   isLoading: boolean;
   resultSet?: ResultSet;
   error?: Error | null;
-};
+} & Partial<QueryStatus>;
 
 type TChartRendererProps = {
   query: Query;
@@ -98,7 +100,7 @@ export default function ChartRenderer({
   onChartRendererReadyChange,
   onQueryStatusChange,
   onRunButtonClick,
-  onQueryChange
+  onQueryChange,
 }: TChartRendererProps) {
   const runButtonRef = useRef<HTMLButtonElement>(null);
   const [slowQuery, setSlowQuery] = useState(false);
@@ -120,29 +122,49 @@ export default function ChartRenderer({
 
   useEffect(() => {
     onQueryChange();
-  }, [areQueriesEqual])
+  }, [areQueriesEqual]);
 
   useEffect(() => {
     setResultSet(false);
   }, [framework]);
 
   useLayoutEffect(() => {
+    let queryStartTime: number;
     window['__cubejsPlayground'] = {
       ...window['__cubejsPlayground'],
       onQueryStart: () => {
+        queryStartTime = Date.now();
         onQueryStatusChange({ isLoading: true });
       },
       onQueryLoad: ({ resultSet, error }: TQueryLoadResult) => {
+        let isAggregated;
+        const timeElapsed = Date.now() - queryStartTime;
         if (resultSet) {
           const { loadResponse } = resultSet.serialize();
 
           setSlowQueryFromCache(Boolean(loadResponse.slowQuery));
           Boolean(loadResponse.slowQuery) && setSlowQuery(false);
           setResultSet(true);
+
+          isAggregated =
+            Object.keys(loadResponse.results[0]?.usedPreAggregations || {})
+              .length > 0;
+
+          event(
+            isAggregated
+              ? 'load_request_success_aggregated:frontend'
+              : 'load_request_success:frontend'
+          );
         }
 
         if (resultSet || error) {
-          onQueryStatusChange({ resultSet, error, isLoading: false });
+          onQueryStatusChange({
+            resultSet,
+            error,
+            isLoading: false,
+            timeElapsed,
+            isAggregated
+          });
         }
       },
       onQueryProgress: (progress) => {